shift payment logic into payment button
